delete wrong files

+++ snr_binned_analysis.py
@@ -205,7 +205,7 @@ def filtering(vars, snrratios=None, time_intervals=None):
             tmp_gti_noflares = f'./data/{source_name_cleaned}/{method}/temp_git_{loop_item}.fits'
             gti_noflares = f'./data/{source_name_cleaned}/{method}/gti_noflares_{loop_item}.fits'                
 
-        if not os.path.exists(gti_noflares and plot_file):
+        if not os.path.exists(gti_noflares):
             ###### Here the BAD TIME intervals (flares) have to be found ######
             f_bin = fits.open(lc)
             bin_data = f_bin[1].data
@@ -1344,25 +1344,30 @@ def run_binned_likelihood(vars, snrratios=None, time_intervals=None, free_params
     return
                    
 def delete_fits_and_xml_files(source_name_cleaned, method):
-    # Define the folders and file types to delete
-    paths_to_delete = [
-        f'./data/{source_name_cleaned}/{method}/srcmap/*.fits',
-        f'./data/{source_name_cleaned}/{method}/models/*.xml',
-        f'./data/{source_name_cleaned}/{method}/ccube/*.fits',
-        f'./data/{source_name_cleaned}/{method}/expmap/*.fits',
-        f'./data/{source_name_cleaned}/{method}/*.fits'
-        f'./data/{source_name_cleaned}/*.fits'
-    ]
-    '''f'./data/{source_name_cleaned}/{method}/ltcube/*.fits','''
-    # Iterate over each path pattern and delete all matching files
-    for path_pattern in paths_to_delete:
-        files = glob.glob(path_pattern)
-        for file in files:
-            try:
-                os.remove(file)
-                #print(f"Deleted: {file}")
-            except OSError as e:
-                print(f"Error deleting file {file}: {e}")
+    # Skip deletion for the protected folder
+    if source_name_cleaned == "4FGLJ0617dot7minus1715":
+        print(f"Skipping deletion for folder: {source_name_cleaned}")
+    else:
+        # Define the folders and file types to delete
+        paths_to_delete = [
+            f'./data/{source_name_cleaned}/{method}/srcmap/*.fits',
+            f'./data/{source_name_cleaned}/{method}/models/*.xml',
+            f'./data/{source_name_cleaned}/{method}/ccube/*.fits',
+            f'./data/{source_name_cleaned}/{method}/expmap/*.fits',
+            f'./data/{source_name_cleaned}/{method}/ltcube/*.fits',
+            f'./data/{source_name_cleaned}/{method}/*.fits',
+            f'./data/{source_name_cleaned}/*.fits'
+        ]
+
+        # Iterate over each path pattern and delete all matching files
+        for path_pattern in paths_to_delete:
+            files = glob.glob(path_pattern)
+            for file in files:
+                try:
+                    os.remove(file)
+                    # print(f"Deleted: {file}")
+                except OSError as e:
+                    print(f"Error deleting file {file}: {e}")
 ##################################################################################
 ##################################################################################
 
@@ -1401,41 +1406,49 @@ def process_line(line):
     vars_lin = (source_name, ra, dec, "LIN", specin, None, None, 100, 1000000)
     source_name_cleaned = source_name.replace(" ", "").replace(".", "dot").replace("+", "plus").replace("-", "minus")
     if not os.path.exists(f"./fit_results/{source_name_cleaned}_fit_data_NONE.fits"):
-        delete_fits_and_xml_files(source_name_cleaned, method = "NONE")
+        
+        '''
         get_gti_bin(vars_none)
         generate_files(vars_none, number_of_bins=7)
         source_maps(vars_none)
         print(source_name)
         run_binned_likelihood(vars_none, free_params="None")
         print(f'Likelihood for non-filtered data done for {source_name}!')
-        
+        '''
+        delete_fits_and_xml_files(source_name_cleaned, method = "NONE")
     else:
         print(f'Likelihood for non-filtered data done for {source_name}!')
-
-    if not os.path.exists(f"./fit_results/{source_name_cleaned}_fit_data_SNR.fits"):
-        filtering(vars_snr, snrratios=snrratios)
-        get_gti_bin(vars_snr, snrratios=snrratios)
-        generate_files(vars_snr, snrratios=snrratios, number_of_bins=7)
-        source_maps(vars_snr, snrratios=snrratios)
-        print(source_name)
-        run_binned_likelihood(vars_snr, snrratios=snrratios, free_params="None")
-        print(f'Likelihood for SNR binned data done for {source_name}!')
-        delete_fits_and_xml_files(source_name_cleaned, method = "SNR")
-        
-    else:
-        print(f'Likelihood for SNR binneddata done for {source_name}!')
-   
+    
     if not os.path.exists(f"./fit_results/{source_name_cleaned}_fit_data_LIN.fits"):
+        '''
         filtering(vars_lin, time_intervals=time_intervals)
         get_gti_bin(vars_lin, time_intervals=time_intervals)
         generate_files(vars_lin, time_intervals=time_intervals, number_of_bins=7)
         source_maps(vars_lin, time_intervals=time_intervals)
         print(source_name)
         run_binned_likelihood(vars_lin, time_intervals=time_intervals, free_params="None")
-        print(f'Likelihood linear binned done for {source_name}!')
+        print(f'Likelihood linear binned done for {source_name}!')'
+        '''
         delete_fits_and_xml_files(source_name_cleaned, method = "LIN")
     else:
         print(f'Likelihood for linear binned data done for {source_name}!')
+
+    if not os.path.exists(f"./fit_results/{source_name_cleaned}_fit_data_SNR.fits"):
+        '''
+        filtering(vars_snr, snrratios=snrratios)
+        get_gti_bin(vars_snr, snrratios=snrratios)
+        generate_files(vars_snr, snrratios=snrratios, number_of_bins=7)
+        source_maps(vars_snr, snrratios=snrratios)
+        print(source_name)
+        run_binned_likelihood(vars_snr, snrratios=snrratios, free_params="None")
+        print(f'Likelihood for SNR binned data done for {source_name}!')'
+        '''
+        delete_fits_and_xml_files(source_name_cleaned, method = "SNR")
+        
+    else:
+        print(f'Likelihood for SNR binneddata done for {source_name}!')
+   
+    
 
 def run_analysis():
     """Main function to use multiprocessing"""
